+++ src/main/java/collections/Collection.java
@@ -1,3 +1,7 @@
+package collections;
+
+import function.Predicate;
+
 import java.util.Iterator;
 
 public interface Collection <E> extends Iterable<E> {
