+++ src/main/java/collections/LinkedSet.java
@@ -1,3 +1,7 @@
+package collections;
+
+import function.Predicate;
+
 import java.util.Iterator;
 
 public class LinkedSet<E> implements Set<E>{
@@ -52,46 +56,6 @@ public class LinkedSet<E> implements Set<E>{
     @Override
     public void clear() {
         linkedList.clear();
-    }
-
-    @Override
-    public Set<E> of() {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e1, E e2) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e1, E e2, E e3) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e1, E e2, E e3, E e4) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e1, E e2, E e3, E e4, E e5) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E e1, E e2, E e3, E e4, E e5, E e6) {
-        return null;
-    }
-
-    @Override
-    public Set<E> of(E... values) {
-        return null;
     }
 
     @Override
