+++ src/main/java/collections/Set.java
@@ -1,0 +1,8 @@
+package collections;
+
+public interface Set<E> extends Collection<E> {
+
+    Set<E> copy();
+
+    Set<E> copyOf(Collection<? extends E> collection);
+}
