+++ src/main/java/function/BinaryConsumer.java
@@ -1,0 +1,6 @@
+package function;
+
+@FunctionalInterface
+public interface BinaryConsumer<T, U> {
+    void accept(T t, U u);
+}
