+++ src/main/java/function/BinaryFunction.java
@@ -1,0 +1,9 @@
+package function;
+
+/*
+Operator, BiOperator, ThreeOperator
+*/
+@FunctionalInterface
+public interface BinaryFunction<T, U, R> {
+    R apply(T t, U u);
+}
