+++ src/main/java/function/BinaryOperator.java
@@ -1,0 +1,6 @@
+package function;
+
+@FunctionalInterface
+public interface BinaryOperator<T> {
+    T apply(T t1, T t2);
+}
