+++ src/main/java/function/BinaryPredicate.java
@@ -1,0 +1,6 @@
+package function;
+
+@FunctionalInterface
+public interface BinaryPredicate<T, U> {
+    boolean test(T t, U u);
+}
