+++ src/main/java/function/Supplier.java
@@ -1,0 +1,6 @@
+package function;
+
+@FunctionalInterface
+public interface Supplier<T> {
+    T get();
+}
