+++ src/main/java/function/TernaryOperator.java
@@ -1,0 +1,5 @@
+package function;
+
+public interface TernaryOperator<T> {
+    T apply(T t1, T t2, T t3);
+}
