+++ src/test/java/ArrayListTest.java
@@ -1,3 +1,4 @@
+import collections.ArrayList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
